Улучшения фильтрации

+++ Tables/ProductsTable.xaml.cs
@@ -70,7 +70,7 @@ namespace ShopManagement.Tables
                 {
                     if (TextBox_PriceBegin.Text.Length > 0)
                     {
-                        FilteredList = FilteredList.Where(Entry => Entry.Price >= Convert.ToInt32(TextBox_PriceBegin.Text)).ToList();
+                        FilteredList = FilteredList.Where(Entry => Entry.Price >= Convert.ToDecimal(TextBox_PriceBegin.Text)).ToList();
                     }
                 }
                 catch { }
@@ -78,7 +78,7 @@ namespace ShopManagement.Tables
                 {
                     if (TextBox_PriceEnd.Text.Length > 0)
                     {
-                        FilteredList = FilteredList.Where(Entry => Entry.Price <= Convert.ToInt32(TextBox_PriceEnd.Text)).ToList();
+                        FilteredList = FilteredList.Where(Entry => Entry.Price <= Convert.ToDecimal(TextBox_PriceEnd.Text)).ToList();
                     }
                 }
                 catch { }
